app-backend/pyproject.toml
[tool.poetry]
name = "django-app"
version = "0.1.0"
description = ""
authors = ["koshkaj <dkoshkeli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
django-redis = "^5.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
